+++ src/rpc/index.ts
@@ -3,8 +3,8 @@ export * from "./json-rpc/request.js";
 export * from "./json-rpc/response.js";
 
 export * from "./client.js";
-export * from "./proxy.js";
 export * from "./request.js";
 export * from "./response.js";
+export * from "./sender.js";
 export * from "./server-client.js";
 export * from "./server.js";